DL-3266 Perform initial load of the Marketing Data Aggregator database

+++ Source/Python/datalabs/etl/orm/load.py
@@ -4,6 +4,7 @@ import io
 import logging
 import math
 
+#import numpy as np
 import pandas
 import sqlalchemy as sa
 
@@ -53,7 +54,7 @@ class ORMLoaderTask(Task):
 
     COLUMN_TYPE_CONVERTERS = {
         'BOOLEAN': lambda x: x.map({'False': False, 'True': True}),
-        'INTEGER': lambda x: x.astype(int, copy=False)
+        'INTEGER': lambda x: x.astype(float).astype('Int64', copy=False)
     }
 
     def run(self):
